+++ core/testsuite/src/main/java/com/blazebit/persistence/testsuite/entity/EmbeddableTestEntity.java
@@ -16,12 +16,18 @@
 
 package com.blazebit.persistence.testsuite.entity;
 
+import javax.persistence.CollectionTable;
 import javax.persistence.DiscriminatorValue;
+import javax.persistence.ElementCollection;
 import javax.persistence.Embedded;
 import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
+import javax.persistence.JoinColumn;
+import javax.persistence.OrderColumn;
 import javax.persistence.Table;
 import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @Table(name = "emb_tst_ent")
@@ -33,6 +39,7 @@ public class EmbeddableTestEntity implements Serializable {
     private EmbeddableTestEntityId id;
     private Long version;
     private EmbeddableTestEntityEmbeddable embeddable = new EmbeddableTestEntityEmbeddable();
+    private List<NameObject> elementCollection4 = new ArrayList<>();
 
     public EmbeddableTestEntity() {
         id = new EmbeddableTestEntityId();
@@ -62,6 +69,22 @@ public class EmbeddableTestEntity implements Serializable {
     
     public void setEmbeddable(EmbeddableTestEntityEmbeddable embeddable) {
         this.embeddable = embeddable;
+    }
+
+    @ElementCollection
+    @OrderColumn(name = "emb_ts_ent_elem_coll4_idx", nullable = false)
+    @CollectionTable(name = "emb_ts_ent_elem_coll4",
+            joinColumns = {
+                    @JoinColumn(name = "elem_coll4_parent_key", referencedColumnName = "test_key"),
+                    @JoinColumn(name = "elem_coll4_parent_value", referencedColumnName = "test_value")
+            }
+    )
+    public List<NameObject> getElementCollection4() {
+        return elementCollection4;
+    }
+
+    public void setElementCollection4(List<NameObject> elementCollection4) {
+        this.elementCollection4 = elementCollection4;
     }
 
 }