+++ core/testsuite/src/test/java/com/blazebit/persistence/testsuite/SelectTest.java
@@ -458,7 +458,7 @@ public class SelectTest extends AbstractCoreTest {
 
         String objectQuery = "SELECT CASE WHEN MIN(d.lastModified) > d.creationDate THEN MIN(d.lastModified) ELSE CURRENT_TIMESTAMP END, owner_1 FROM Document d "
                 + "JOIN d.owner owner_1 "
-                + "GROUP BY " + groupBy("d.creationDate", "owner_1.age", "owner_1.defaultLanguage", "owner_1.friend", "owner_1.id", "owner_1.name", "owner_1.nameObject", "owner_1.partnerDocument", renderNullPrecedenceGroupBy("d.id"))
+                + "GROUP BY " + groupBy("d.creationDate", "owner_1.age", "owner_1.defaultLanguage", "owner_1.friend", "owner_1.id", "owner_1.name", "owner_1.nameObject.intIdEntity.id", "owner_1.nameObject.primaryName", "owner_1.nameObject.secondaryName", "owner_1.partnerDocument", renderNullPrecedenceGroupBy("d.id"))
                 + " ORDER BY d.id DESC";
 
         assertEquals(objectQuery, cb.getQueryString());
@@ -477,7 +477,7 @@ public class SelectTest extends AbstractCoreTest {
         String countQuery = "SELECT " + countPaginated("d.id", false) + " FROM Document d";
         String objectQuery = "SELECT CASE WHEN MIN(d.lastModified) > d.creationDate THEN MIN(d.lastModified) ELSE CURRENT_TIMESTAMP END, owner_1 FROM Document d "
                 + "JOIN d.owner owner_1 "
-                + "GROUP BY " + groupBy("d.creationDate", "owner_1.age", "owner_1.defaultLanguage", "owner_1.friend", "owner_1.id", "owner_1.name", "owner_1.nameObject", "owner_1.partnerDocument", renderNullPrecedenceGroupBy("d.id"))
+                + "GROUP BY " + groupBy("d.creationDate", "owner_1.age", "owner_1.defaultLanguage", "owner_1.friend", "owner_1.id", "owner_1.name", "owner_1.nameObject.intIdEntity.id", "owner_1.nameObject.primaryName", "owner_1.nameObject.secondaryName", "owner_1.partnerDocument", renderNullPrecedenceGroupBy("d.id"))
                 + " ORDER BY d.id DESC";
 
         assertEquals(countQuery, cb.getPageCountQueryString());