+++ integration/datanucleus/src/main/java/com/blazebit/persistence/integration/datanucleus/DataNucleusJpaProvider.java
@@ -469,7 +469,7 @@ public class DataNucleusJpaProvider implements JpaProvider {
     }
 
     @Override
-    public boolean needsElementCollectionIdCutoff() {
+    public boolean needsElementCollectionIdCutoffForCompositeIdOwner() {
         return false;
     }